test: move governance mock
